feat(cloudevents-server): support sink events and produce for tibuild
