+++ cloudevents-server/main.go
@@ -4,13 +4,11 @@ import (
 	"flag"
 	"net/http"
 
-	cloudevents "github.com/cloudevents/sdk-go/v2"
 	"github.com/gin-gonic/gin"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
 	"github.com/PingCAP-QE/ee-apps/cloudevents-server/pkg/config"
-	"github.com/PingCAP-QE/ee-apps/cloudevents-server/pkg/events"
 )
 
 func main() {
@@ -38,7 +36,7 @@ func main() {
 	case gin.DebugMode:
 		zerolog.SetGlobalLevel(zerolog.DebugLevel)
 	default:
-		zerolog.SetGlobalLevel(zerolog.DebugLevel)
+		zerolog.SetGlobalLevel(zerolog.InfoLevel)
 	}
 
 	gin.SetMode(ginMode)
@@ -56,34 +54,5 @@ func main() {
 func setRouters(r gin.IRoutes, cfg *config.Config) {
 	r.GET("/", indexHandler)
 	r.GET("/healthz", healthzHandler)
-	r.POST("/events", eventsHandler(cfg))
-}
-
-func indexHandler(c *gin.Context) {
-	c.JSON(http.StatusOK, "Welcome to CloudEvents")
-}
-
-func healthzHandler(c *gin.Context) {
-	c.String(http.StatusOK, "OK")
-}
-
-func eventsHandler(cfg *config.Config) gin.HandlerFunc {
-	p, err := cloudevents.NewHTTP()
-	if err != nil {
-		log.Fatal().Err(err).Msg("Failed to create protocol")
-	}
-
-	handler, err := events.NewEventsHandler(cfg)
-	if err != nil {
-		log.Fatal().Err(err).Msg("failed to create cloudevents handler")
-	}
-
-	h, err := cloudevents.NewHTTPReceiveHandler(nil, p, handler.Handle)
-	if err != nil {
-		log.Fatal().Err(err).Msg("failed to create handler")
-	}
-
-	return func(c *gin.Context) {
-		h.ServeHTTP(c.Writer, c.Request)
-	}
+	r.POST("/events", newEventsHandlerFunc(cfg))
 }
